Moved dashboard save modals to components folder
